# pkg/api/api.swagger.json
{
  "consumes": [
    "application/grpc",
    "application/json"
  ],
  "produces": [
    "application/grpc",
    "application/json"
  ],
  "schemes": [
    "http",
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "API for simple user authorization",
    "title": "SSO Test Service API",
    "version": "1.0"
  },
  "paths": {
    "/api/AccessV1/Check": {
      "post": {
        "description": "Проверь сервис на возможность использование",
        "tags": [
          "AccessV1"
        ],
        "summary": "Проверить сервис",
        "operationId": "AccessV1_Check",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/access_v1CheckRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        }
      }
    },
    "/api/AuthV1/Login": {
      "post": {
        "description": "Аутентификация пользователя",
        "tags": [
          "AuthV1"
        ],
        "summary": "Аутентификация пользователя",
        "operationId": "AuthV1_Login",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth_v1LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/auth_v1LoginResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        }
      }
    },
    "/api/AuthV1/Logout": {
      "get": {
        "description": "Завершить сессию пользователя",
        "tags": [
          "AuthV1"
        ],
        "summary": "Выйти из пользователя",
        "operationId": "AuthV1_Logout",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "object"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        }
      }
    },
    "/api/AuthV1/UpdateToken": {
      "post": {
        "description": "Обновление токенов",
        "tags": [
          "AuthV1"
        ],
        "summary": "Обновление токенов",
        "operationId": "AuthV1_UpdateToken",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/auth_v1UpdateTokenRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/auth_v1UpdateTokenResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        }
      }
    },
    "/api/CommonV1/GetTime": {
      "get": {
        "description": "Returns the current server time",
        "tags": [
          "CommonV1"
        ],
        "summary": "Get current time",
        "operationId": "CommonV1_GetTime",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/common_v1TimeResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        }
      }
    },
    "/api/CommonV1/LongOperation": {
      "post": {
        "description": "Initiates a long operation and returns streaming updates about its progress",
        "tags": [
          "CommonV1"
        ],
        "summary": "Perform a long operation",
        "operationId": "CommonV1_LongOperation",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/common_v1LongOperationRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A successful response.(streaming responses)",
            "schema": {
              "type": "object",
              "title": "Stream result of common_v1LongOperationResponse",
              "properties": {
                "error": {
                  "$ref": "#/definitions/googlerpcStatus"
                },
                "result": {
                  "$ref": "#/definitions/common_v1LongOperationResponse"
                }
              }
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "access_v1CheckRequest": {
      "type": "object",
      "properties": {
        "endpointAddress": {
          "description": "Endpoint address",
          "type": "string",
          "title": "endpoint_address - адрес ресурса, который требуется проверить",
          "example": "Hello, world!"
        }
      }
    },
    "auth_v1LoginRequest": {
      "type": "object",
      "properties": {
        "password": {
          "description": "Password",
          "type": "string",
          "example": "qwerty"
        },
        "username": {
          "description": "Username",
          "type": "string",
          "example": "root"
        }
      }
    },
    "auth_v1LoginResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "description": "Access token",
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        },
        "refreshToken": {
          "description": "Refresh token",
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjE2Nzg5MCIsIm5hbWUiOiJKb2huIERvZSIsImlhdCI6MTUxNjIzOTAyMn0.wZ3tAPWpifvkvKEizMMqR8B2QJeD2Q1nFWBRPdE1nrg"
        }
      }
    },
    "auth_v1UpdateTokenRequest": {
      "type": "object",
      "properties": {
        "refreshToken": {
          "description": "Refresh token",
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjE2Nzg5MCIsIm5hbWUiOiJKb2huIERvZSIsImlhdCI6MTUxNjIzOTAyMn0.wZ3tAPWpifvkvKEizMMqR8B2QJeD2Q1nFWBRPdE1nrg"
        }
      }
    },
    "auth_v1UpdateTokenResponse": {
      "type": "object",
      "properties": {
        "accessToken": {
          "description": "Access token",
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        },
        "refreshToken": {
          "description": "Refresh token",
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjE2Nzg5MCIsIm5hbWUiOiJKb2huIERvZSIsImlhdCI6MTUxNjIzOTAyMn0.wZ3tAPWpifvkvKEizMMqR8B2QJeD2Q1nFWBRPdE1nrg"
        }
      }
    },
    "common_v1LongOperationRequest": {
      "type": "object"
    },
    "common_v1LongOperationResponse": {
      "type": "object",
      "properties": {
        "message": {
          "description": "Progress message",
          "type": "string"
        },
        "progress": {
          "description": "Completion percentage (0-100)",
          "type": "integer",
          "format": "int32"
        },
        "result": {
          "description": "Operation result (if completed)",
          "type": "string"
        },
        "status": {
          "description": "Operation status",
          "$ref": "#/definitions/common_v1LongOperationResponseStatus"
        }
      }
    },
    "common_v1LongOperationResponseStatus": {
      "type": "string",
      "default": "IN_PROGRESS",
      "enum": [
        "IN_PROGRESS",
        "COMPLETED",
        "FAILED"
      ]
    },
    "common_v1TimeResponse": {
      "type": "object",
      "properties": {
        "time": {
          "description": "Current time",
          "type": "string",
          "format": "date-time",
          "example": "2023-04-01T12:00:00Z"
        }
      }
    },
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        },
        "message": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        },
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "X-API-Key",
      "in": "header"
    }
  },
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "tags": [
    {
      "name": "CommonV1"
    },
    {
      "name": "AuthV1"
    },
    {
      "name": "AccessV1"
    }
  ]
}